created graphics for tsp

+++ travelin-salesman/src/main/java/com/bbw/ch/Node.java
@@ -1,29 +1,29 @@
 package com.bbw.ch;
 
 public class Node {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
     private String name;
 
-    public Node(int x, int y, String name) {
+    public Node(double x, double y, String name) {
         this.x = x;
         this.y = y;
         this.name = name;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
@@ -36,8 +36,8 @@ public class Node {
     }
 
     public double calcWay(Node node) {
-        int x = node.getX() > this.getX() ? node.getX() - this.getX() : this.getX() - node.getX();
-        int y = node.getY() > this.getY() ? node.getY() - this.getY() : this.getY() - node.getY();
+        double x = node.getX() > this.getX() ? node.getX() - this.getX() : this.getX() - node.getX();
+        double y = node.getY() > this.getY() ? node.getY() - this.getY() : this.getY() - node.getY();
 
         return Math.sqrt(Math.pow(x, 2) + Math.pow(y, 2));
     }
